+++ src/Controllers/GlobalData/MeasurementUnitController.php
@@ -5,11 +5,12 @@ namespace JtlWooCommerceConnector\Controllers\GlobalData;
 use Jtl\Connector\Core\Model\Identity;
 use Jtl\Connector\Core\Model\MeasurementUnit as MeasurementUnitModel;
 use Jtl\Connector\Core\Model\MeasurementUnitI18n;
-use JtlWooCommerceConnector\Controllers\AbstractController;
+use JtlWooCommerceConnector\Controllers\AbstractBaseController;
+use JtlWooCommerceConnector\Integrations\Plugins\Wpml\WpmlGermanMarket;
 use JtlWooCommerceConnector\Utilities\Germanized;
 use JtlWooCommerceConnector\Utilities\SqlHelper;
 
-class MeasurementUnitController extends AbstractController
+class MeasurementUnitController extends AbstractBaseController
 {
     /**
      * @return array
@@ -19,6 +20,11 @@ class MeasurementUnitController extends AbstractController
     {
         $measurementUnits = [];
 
+        $defaultLanguage = $this->util->getWooCommerceLanguage();
+        if ($this->wpml->canBeUsed()) {
+            $defaultLanguage = $this->wpml->convertLanguageToWawi($this->wpml->getDefaultLanguage());
+        }
+
         $result = $this->db->query(SqlHelper::globalDataGermanizedMeasurementUnitPull());
 
         foreach ((array)$result as $row) {
@@ -26,11 +32,11 @@ class MeasurementUnitController extends AbstractController
                 ->setId(new Identity($row['id']))
                 ->setCode((new Germanized())->parseUnit($row['code']))
                 ->setDisplayCode($row['code'])
-                ->setI18ns(...[
+                ->setI18ns(
                     (new MeasurementUnitI18n())
                         ->setName($row['code'])
-                        ->setLanguageISO($this->util->getWooCommerceLanguage()),
-                ]);
+                        ->setLanguageISO($defaultLanguage)
+                );
         }
 
         return $measurementUnits;
@@ -39,6 +45,7 @@ class MeasurementUnitController extends AbstractController
     /**
      * @return array
      * @throws \InvalidArgumentException
+     * @throws \Exception
      */
     public function pullGermanMarketData(): array
     {
@@ -59,15 +66,27 @@ class MeasurementUnitController extends AbstractController
         )));
 
         foreach ($values as $unit) {
-            $measurementUnits[] = (new MeasurementUnitModel())
+            $measurementUnit = (new MeasurementUnitModel())
                 ->setId(new Identity($unit['term_id']))
                 ->setCode($unit['name'])
                 ->setDisplayCode($unit['name'])
-                ->setI18ns(...[
+                ->setI18ns(
                     (new MeasurementUnitI18n())
                         ->setName($unit['description'])
-                        ->setLanguageISO($this->util->getWooCommerceLanguage()),
-                ]);
+                        ->setLanguageISO($this->util->getWooCommerceLanguage())
+                );
+
+            if ($this->wpml->canBeUsed()) {
+                $translations = $this->wpml
+                    ->getComponent(WpmlGermanMarket::class)
+                    ->getMeasurementUnitsTranslations($unit['term_taxonomy_id'], $specific['attribute_name']);
+
+                foreach ($translations as $translation) {
+                    $measurementUnit->addI18n($translation);
+                }
+            }
+
+            $measurementUnits[] = $measurementUnit;
         }
 
         return $measurementUnits;