+++ src/Controllers/Product/ProductAttrController.php
@@ -3,7 +3,6 @@
 namespace JtlWooCommerceConnector\Controllers\Product;
 
 use Jtl\Connector\Core\Model\Identity;
-use InvalidArgumentException;
 use Jtl\Connector\Core\Exception\TranslatableAttributeException;
 use Jtl\Connector\Core\Model\Product as ProductModel;
 use Jtl\Connector\Core\Model\TranslatableAttribute as ProductAttrModel;
@@ -11,8 +10,6 @@ use Jtl\Connector\Core\Model\TranslatableAttributeI18n as ProductAttrI18nModel;
 use JtlWooCommerceConnector\Controllers\AbstractBaseController;
 use JtlWooCommerceConnector\Utilities\Config;
 use JtlWooCommerceConnector\Utilities\SupportedPlugins;
-use WC_Product;
-use WC_Product_Attribute;
 
 class ProductAttrController extends AbstractBaseController
 {
@@ -68,11 +65,16 @@ class ProductAttrController extends AbstractBaseController
         $suppressShippingNotice = false;
         $variationPreselect     = [];
 
+        //GERMANIZED PRO
+        $food = false;
+
         /** @var  ProductAttrModel $pushedAttribute */
         foreach ($pushedAttributes as $key => $pushedAttribute) {
             foreach ($pushedAttribute->getI18ns() as $i18n) {
                 if (!$this->util->isWooCommerceLanguage($i18n->getLanguageISO())) {
-                    continue;
+                    if (!\in_array($i18n->getLanguageIso(), \array_keys($this->wpml->getActiveLanguages()))) {
+                        continue;
+                    }
                 }
 
                 $attrName = \strtolower(\trim($i18n->getName()));
@@ -96,6 +98,14 @@ class ProductAttrController extends AbstractBaseController
                         }
                         if ($i18n->getName() === ProductVaSpeAttrHandlerController::GZD_MIN_AGE) {
                             $this->addOrUpdateMetaField($productId, '_min_age', $i18n->getValue());
+                        }
+                    }
+
+                    if (SupportedPlugins::isActive(SupportedPlugins::PLUGIN_WOOCOMMERCE_GERMANIZEDPRO)) {
+                        if ($i18n->getName() === ProductVaSpeAttrHandlerController::GZD_IS_FOOD) {
+                            $value = $this->util->isTrue($i18n->getValue()) ? 'yes' : 'no';
+                            $this->addOrUpdateMetaField($productId, '_is_food', $value);
+                            $food = true;
                         }
                     }
 
@@ -259,6 +269,10 @@ class ProductAttrController extends AbstractBaseController
             );
         }
 
+        if (SupportedPlugins::isActive(SupportedPlugins::PLUGIN_WOOCOMMERCE_GERMANIZEDPRO) && !$food) {
+            $this->addOrUpdateMetaField($productId, '_is_food', 'no');
+        }
+
         if (!$payable) {
             $wcProduct = \wc_get_product($productId);
             $wcProduct->set_status('publish');
@@ -281,6 +295,8 @@ class ProductAttrController extends AbstractBaseController
             }
         }
 
+        $sentCustomProperties = [];
+
         /** @var ProductAttrModel $attribute */
         foreach ($pushedAttributes as $attribute) {
             if (
@@ -295,10 +311,14 @@ class ProductAttrController extends AbstractBaseController
                     continue;
                 }
 
+                $sentCustomProperties[] = $attribute->getName();
+
                 $this->saveAttribute($attribute, $i18n, $attributesFilteredVariationsAndSpecifics);
                 break;
             }
         }
+
+        $this->deleteRemovedCustomProperties($productId, $sentCustomProperties);
 
         return $attributesFilteredVariationsAndSpecifics;
     }
@@ -403,6 +423,38 @@ class ProductAttrController extends AbstractBaseController
             'is_variation' => 0,
             'is_taxonomy' => 0,
         ];
+    }
+
+    private function deleteRemovedCustomProperties($productId, $sentCustomProperties): void
+    {
+        global $wpdb;
+
+        $query = \sprintf(
+            "
+            SELECT meta_value
+            FROM {$wpdb->postmeta}
+            WHERE post_id = %d
+            AND meta_key = '_product_attributes'",
+            $productId
+        );
+
+        $existingProperties    = $this->db->query($query);
+        $existingProperties    = \unserialize($existingProperties[0]['meta_value']);
+        $existingPropertyNames = [];
+
+        foreach ($existingProperties as $property) {
+            $existingPropertyNames[] = $property['name'];
+        }
+
+        $missingProperties = \array_diff($existingPropertyNames, $sentCustomProperties);
+
+        if ($missingProperties) {
+            foreach ($missingProperties as $missingKey) {
+                unset($existingProperties[\str_replace(' ', '-', \strtolower($missingKey))]);
+            }
+
+            \update_post_meta($productId, '_product_attributes', $existingProperties);
+        }
     }
 
     /**