+++ CHANGELOG.md
@@ -3,8 +3,16 @@
 This is the changelog of our "JTL WooCommerce Connector".
 
 ## Unreleased
-- CO-2523 - get build version from config
+- CO-2514 - prevent duplicate product images 
+- CO-2394 - delete removed custom attributes
+- CO-2460 - implement compatibility with germanized pro food 
+- CO-2582 - merge wpml with wc connector
 - CO-2440 - support advanced custom fields plugin for products
+
+## 1.42.2 _2024-02-14_
+- CO-2523 - get build version from config
+- CO-2527 - pass rank math data as string
+- CO-2536 - fix hpos payment
 
 ## 1.42.1 _2024-01-19_
 - CO-2528 - update operand type